# openapi.json
openapi: 3.0.3
info:
  title: University SIS API
  version: 1.0.0
  description: Academic records, scheduling, and enrollment endpoints.
paths:
...
  /api/academics/cohorts/add-group/:
    post:
      operationId: academics_cohorts_add_group_create
      description: |-
        Handles all cohort-level operations (listing, creation, scheduling,
        adding study groups, and deletion). Cohorts are virtual entities
        represented by (discipline, term, year_level) triples.
      summary: Add a study group to an existing cohort
      tags:
      - academics
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStudyGroupRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AddStudyGroupRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AddStudyGroupRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddStudyGroupResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/academics/cohorts/bulk-cohort/:
    post:
      operationId: academics_cohorts_bulk_cohort_create
      description: Atomically creates all StudyGroup and CourseClass instances for
        a new cohort.
      summary: Bulk create a cohort
      tags:
      - academics
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortBulkCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CohortBulkCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CohortBulkCreate'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkCreateCohortResponse'
          description: ''
  /api/academics/cohorts/cohorts/:
    get:
      operationId: academics_cohorts_cohorts_list
      description: Retrieves a list of cohorts grouped by discipline, term, and year
        level.
      summary: List all cohorts
      parameters:
      - in: query
        name: discipline
        schema:
          type: integer
        description: Alias for discipline_id
      - in: query
        name: discipline_id
        schema:
          type: integer
        description: Filter by discipline ID
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: term_status
        schema:
          type: string
          default: active
        description: Filter by term status (active, past, all)
      tags:
      - academics
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCohortReadList'
          description: ''
    delete:
      operationId: academics_cohorts_cohorts_destroy
      description: Deletes all study groups and associated course classes for the
        specified cohort.
      summary: Delete a cohort
      parameters:
      - in: query
        name: discipline_id
        schema:
          type: integer
        required: true
      - in: query
        name: term_id
        schema:
          type: integer
        required: true
      - in: query
        name: year_level
        schema:
          type: integer
          maximum: 4
          minimum: 1
        required: true
      tags:
      - academics
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
        '404':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/academics/cohorts/schedule-cohort/:
    post:
      operationId: academics_cohorts_schedule_cohort_create
      description: Runs the CP-SAT solver to generate a conflict-free timetable for
        the cohort.
      summary: Schedule a cohort
      tags:
      - academics
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleCohortRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ScheduleCohortRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ScheduleCohortRequest'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleCohortResponse'
          description: ''
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleCohortResponse'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RescheduleConfirmationResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
  schemas:
    AddStudyGroupRequest:
      type: object
      properties:
        discipline_id:
          type: integer
        term_id:
          type: integer
        year_level:
          type: integer
          maximum: 4
          minimum: 1
        number:
          type: integer
          minimum: 1
        capacity:
          type: integer
          minimum: 1
          default: 50
      required:
      - discipline_id
      - number
      - term_id
      - year_level
    AddStudyGroupResponse:
      type: object
      properties:
        message:
          type: string
        id:
          type: integer
      required:
      - id
      - message
...
    BulkCreateCohortResponse:
      type: object
      properties:
        message:
          type: string
        groups_created:
          type: integer
        classes_created:
          type: integer
        group_ids:
          type: array
          items:
            type: integer
        class_ids:
          type: array
          items:
            type: integer
      required:
      - class_ids
      - classes_created
      - group_ids
      - groups_created
      - message
    CohortBulkCreate:
      type: object
      properties:
        discipline_id:
          type: integer
        term_id:
          type: integer
        year_level:
          type: integer
          maximum: 4
          minimum: 1
        groups:
          type: array
          items:
            $ref: '#/components/schemas/StudyGroupInput'
        courses:
          type: array
          items:
            type: integer
        coordinators:
          type: object
          additionalProperties:
            type: integer
            nullable: true
      required:
      - courses
      - discipline_id
      - groups
      - term_id
      - year_level
    CohortCourseClassRead:
      type: object
      description: Serializes a CourseClass enriched with its group number and coordinator
        name.
      properties:
        id:
          type: integer
        course:
          $ref: '#/components/schemas/_CourseInline'
        group_number:
          type: integer
        coordinator:
          allOf:
          - $ref: '#/components/schemas/_CoordinatorInline'
          nullable: true
      required:
      - coordinator
      - course
      - group_number
      - id
    CohortGroupRead:
      type: object
      description: Serializes a single StudyGroup as a lightweight group descriptor.
      properties:
        id:
          type: integer
        number:
          type: integer
        capacity:
          type: integer
      required:
      - capacity
      - id
      - number
    CohortRead:
      type: object
      description: |-
        Virtual serializer for a Cohort — a logical grouping of StudyGroups sharing
        the same (discipline, term, year_level) triple.

        Instances are plain dicts built by CohortViewSet.cohorts() rather than
        ORM model instances.
      properties:
        id:
          type: string
          readOnly: true
        discipline:
          $ref: '#/components/schemas/Discipline'
        term:
          $ref: '#/components/schemas/Term'
        year_level:
          type: integer
        groups:
          type: array
          items:
            $ref: '#/components/schemas/CohortGroupRead'
        course_classes:
          type: array
          items:
            $ref: '#/components/schemas/CohortCourseClassRead'
        schedule_status:
          $ref: '#/components/schemas/ScheduleStatusEnum'
      required:
      - course_classes
      - discipline
      - groups
      - id
      - schedule_status
      - term
      - year_level
...
    Department:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        code:
          type: string
          maxLength: 10
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - code
      - created_at
      - id
      - name
      - updated_at
    Discipline:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        code:
          type: string
          maxLength: 20
        department:
          allOf:
          - $ref: '#/components/schemas/Department'
          readOnly: true
        program_type:
          $ref: '#/components/schemas/ProgramTypeEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - code
      - created_at
      - department
      - id
      - name
      - program_type
      - updated_at
...
    PaginatedCohortReadList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/CohortRead'
...
    ProgramTypeEnum:
      enum:
      - GSP
      - SSP
      type: string
      description: |-
        * `GSP` - General Specialization Program
        * `SSP` - Specific Specialization Program
...
    RescheduleConfirmationResponse:
      type: object
      properties:
        detail:
          type: string
        requires_confirmation:
          type: boolean
        stale_session_count:
          type: integer
        affected_enrollment_count:
          type: integer
      required:
      - affected_enrollment_count
      - detail
      - requires_confirmation
      - stale_session_count
...
    ScheduleCohortRequest:
      type: object
      properties:
        discipline_id:
          type: integer
        term_id:
          type: integer
        year_level:
          type: integer
          maximum: 4
          minimum: 1
        time_limit:
          type: integer
          maximum: 300
          minimum: 5
          default: 60
        dry_run:
          type: boolean
          default: false
        force:
          type: boolean
          default: false
      required:
      - discipline_id
      - term_id
      - year_level
    ScheduleCohortResponse:
      type: object
      properties:
        status:
          type: string
        sessions_created:
          type: integer
        course_classes_scheduled:
          type: integer
        solve_time_seconds:
          type: number
          format: double
        dry_run:
          type: boolean
      required:
      - course_classes_scheduled
      - dry_run
      - sessions_created
      - solve_time_seconds
      - status
...
    ScheduleStatusEnum:
      enum:
      - unscheduled
      - needs_reschedule
      - scheduled
      type: string
      description: |-
        * `unscheduled` - unscheduled
        * `needs_reschedule` - needs_reschedule
        * `scheduled` - scheduled
...
    StudyGroupInput:
      type: object
      properties:
        number:
          type: integer
          minimum: 1
        capacity:
          type: integer
          minimum: 1
          default: 50
      required:
      - number
...
    Term:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        is_active:
          type: boolean
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - end_date
      - id
      - is_active
      - name
      - start_date
      - updated_at
...
    _CoordinatorInline:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          readOnly: true
      required:
      - id
      - name
    _CourseInline:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        title:
          type: string
      required:
      - code
      - id
      - title